build: Migrate to pdm-backend
Migrate from the deprecated pdm-pep517 backend to pdm-backend that
superseded it.  This required moving the build-specific keys from
tool.pdm table to tool.pdm.backend, as well as modernizing the SCM
version usage.  The artifacts produced after this change are equivalent,
except that their names are normalized to "mkdocs_autorefs" as required
by modern PyPA standards.

The migration guide is at:
https://pdm-backend.fming.dev/migration/

It does not mention SCM support, probably because the "new" metadata
is already supported by newer pdm-pep517 versions.  The docs are at:
https://pdm-backend.fming.dev/metadata/#read-from-scm-tag-supporting-git-and-hg