feat: Separate URLs in two groups, primary and secondary
This will be useful in mkdocstrings, where we want to register URLs for all aliases of a rendered object's identifier early in the process, so that we can drop the fallback mechanism in autorefs.

Primary URLs will take precedence when resolving cross-references, to avoid logging warnings about multiple URLs found.

For example:

- Object `a.b.c.d` has aliases `a.b.d` and `a.d`
- Object `a.b.c.d` is rendered.
- We register `a.b.c.d` -> page#a.b.c.d as primary
- We register `a.b.d` -> page#a.b.c.d as secondary
- We register `a.d` -> page#a.b.c.d as secondary
- Later, if `a.b.d` or `a.d` are rendered, we will register primary and secondary URLs the same way
- This way we are sure that each of `a.b.c.d`, `a.b.d` or `a.d` will link to their primary URL, if any, or their secondary URL, accordingly

Related-to-issue-61: https://github.com/mkdocstrings/autorefs/issues/61