feat: Add `link_titles` option and adapt related logic
This change adds a `link_titles` option that defaults to `"auto"`. In automatic mode, link titles are either:

- always set if Material for MkDocs and its instant preview feature aren't detected
- only set on external links otherwise (since instant preview are not supported on external links anyway)

The option also accepts the `True` and `False`, for always/never setting titles, respectively.

An update to the title logic accompanies this change in order to make use of recorded heading titles (a change brought two commit ago):

- optional cross-references will use the original title, and optionally append the identifier if it doesn't already appear in the title
- mandatory cross-references will use either the original title if there's one, or no title at all

This is because optional cross-refs are almost exclusively created by mkdocstrings handlers, and therefore displaying the identifier (full qualified name of objects) is useful when hovering on a link. Manual cross-references on the other hand can often be references to text sections, and should never display the section anchor. The limitation being that manual cross-references to API objects won't show the identifier. We could consider using an additional attribute (other than `optional`) to label cross-refs as "API objects" or not, though users would still have to annotate their manual cross-refs with such an attribute to enjoy the appended identifier to the title.

Issue-33: https://github.com/mkdocstrings/autorefs/issues/33
Issue-62: https://github.com/mkdocstrings/autorefs/issues/62